@@ -12,56 +12,59 @@
     "url": "git@github.com:negative0/rclone-webui-react.git"
   },
   "dependencies": {
-    "@coreui/coreui": "^2.1.16",
-    "@coreui/coreui-plugin-chartjs-custom-tooltips": "^1.3.1",
-    "@coreui/icons": "^0.3.0",
-    "@coreui/react": "^2.5.4",
-    "axios": "^0.21.2",
-    "bootstrap": "^4.6.0",
-    "chart.js": "^2.9.4",
-    "classnames": "^2.3.1",
-    "core-js": "^3.15.2",
-    "flag-icon-css": "^3.5.0",
-    "font-awesome": "^4.7.0",
-    "jquery": "^3.6.0",
-    "lodash": "^4.17.21",
-    "prop-types": "^15.7.2",
-    "rclone-api": "^1.0.10",
-    "react": "^16.14.0",
-    "react-app-polyfill": "^2.0.0",
-    "react-autosuggest": "^10.1.0",
-    "react-awesome-player": "^1.1.0",
-    "react-chartjs-2": "^2.11.2",
+    "@coreui/react": "^4.11.1",
+    "react-router-dom": "^5.3.4",
+    "redux-thunk": "^2.4.2",
+    "reactstrap": "^8.9.0",
     "react-dnd": "^11.1.3",
     "react-dnd-html5-backend": "^11.1.3",
-    "react-dom": "^16.14.0",
-    "react-in-viewport": "^1.0.0-alpha.19",
-    "react-redux": "^7.2.4",
+    "flag-icons": "^7.1.0", 
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@coreui/coreui": "^5.1.2",
+    "@coreui/coreui-plugin-chartjs-custom-tooltips": "^1.3.1",
+    "@coreui/icons": "^3.0.1",
+    "@coreui/react-chartjs": "^3.0.0",
+    "axios": "^1.6.7",
+    "bootstrap": "^5.3.3",
+    "chart.js": "^4.4.1",
+    "classnames": "^2.5.1",
+    "core-js": "^3.35.1",
+    "flag-icon-css": "^4.1.7",
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.7.1",
+    "lodash": "^4.17.21",
+    "prop-types": "^15.8.1",
+    "rclone-api": "^1.0.10",
+    "react": "^18.2.0",
+    "react-app-polyfill": "^3.0.0",
+    "react-autosuggest": "^10.1.0",
+    "react-awesome-player": "^1.0.7",
+    "react-chartjs-2": "^5.2.0",
+    "react-dom": "^18.2.0",
+    "react-in-viewport": "^1.0.0-alpha.30",
+    "react-redux": "^9.1.0",
     "react-router-config": "^5.1.1",
-    "react-router-dom": "^5.2.0",
-    "react-test-renderer": "^16.14.0",
-    "react-toastify": "^7.0.4",
-    "reactstrap": "^8.9.0",
-    "redux": "^4.1.0",
-    "redux-thunk": "^2.3.0",
-    "reselect": "^4.0.0",
+    "react-test-renderer": "^18.2.0",
+    "react-toastify": "^10.0.4",
+    "redux": "^5.0.1",
+    "reselect": "^5.1.0",
+    "sass": "^1.70.0",
     "simple-line-icons": "^2.5.5",
-    "typescript": "^4.3.5",
-    "sass": "^1.32.8"
+    "typescript": "^5.3.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.6",
+    "@babel/core": "^7.23.9",
     "check-prop-types": "^1.1.2",
     "coveralls": "^3.1.1",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
+    "enzyme-adapter-react-16": "^1.15.7",
     "enzyme-to-json": "^3.6.2",
     "fetch-mock": "^9.11.0",
-    "react-dnd-test-backend": "^11.1.3",
-    "react-dnd-test-utils": "^11.1.3",
-    "react-scripts": "^4.0.3",
+    "react-dnd-test-backend": "^16.0.1",
+    "react-dnd-test-utils": "^16.0.1",
+    "react-scripts": "^5.0.1",
     "redux-mock-store": "^1.5.4",
-    "sass-loader": "^10.0.5"
+    "sass-loader": "^14.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -85,8 +88,8 @@
     "not op_mini all"
   ],
   "resolutions": {
-    "browserslist": "^4.16.6",
-    "caniuse-lite": "^1.0.30001240"
+    "browserslist": "^4.22.3",
+    "caniuse-lite": "^1.0.30001576"
   },
   "jest": {
     "collectCoverageFrom": [
