@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {HashRouter, Route, Switch} from 'react-router-dom';
+import {HashRouter, Route, Routes} from 'react-router-dom';
 // import { renderRoutes } from 'react-router-config';
 import './App.scss';
 import {ToastContainer} from "react-toastify";
@@ -26,14 +26,14 @@ class App extends Component {
                     <ToastContainer/>
                     <HashRouter>
                         <React.Suspense fallback={loading()}>
-                            <Switch>
+                            <Routes>
                                 <Route exact path="/login" name="Login Page" render={props => <Login {...props}/>}/>
                                 <Route exact path="/register" name="Register Page"
                                        render={props => <Register {...props}/>}/>
                                 <Route exact path="/404" name="Page 404" render={props => <Page404 {...props}/>}/>
                                 <Route exact path="/500" name="Page 500" render={props => <Page500 {...props}/>}/>
                                 <Route path="/" name="Home" render={props => <DefaultLayout {...props}/>}/>
-                            </Switch>
+                            </Routes>
                         </React.Suspense>
                     </HashRouter>
                 </ErrorBoundary>
