@@ -1,28 +1,26 @@
-import React, {Component, Suspense} from 'react';
-import {Redirect, Route, Switch} from 'react-router-dom';
-import {Container} from 'reactstrap';
-import {getVersion} from "../../actions/versionActions";
+import React, { Component, Suspense } from 'react';
+import { useNavigate, Routes, Route, Navigate, useLocation } from 'react-router-dom';
+import { Container } from 'reactstrap';
+import { getVersion } from "../../actions/versionActions";
 
 import {
-    AppBreadcrumb,
-    AppFooter,
-    AppHeader,
-    AppSidebar,
-    AppSidebarFooter,
-    AppSidebarForm,
-    AppSidebarHeader,
-    AppSidebarMinimizer,
-    AppSidebarNav,
+    CBreadcrumb,
+    CFooter,
+    CHeader,
+    CSidebar,
+    CSidebarBrand,
+    CSidebarHeader,
+    CSidebarNav,
 } from '@coreui/react';
+
 // sidebar nav config
 import navigation from '../../_nav';
 // routes config
 import routes from '../../routes';
-import {connect} from "react-redux";
-import {AUTH_KEY, LOGIN_TOKEN} from "../../utils/Constants";
+import { connect } from "react-redux";
+import { AUTH_KEY, LOGIN_TOKEN } from "../../utils/Constants";
 import ErrorBoundary from "../../ErrorHandling/ErrorBoundary";
 
-// const DefaultAside = React.lazy(() => import('./DefaultAside'));
 const DefaultFooter = React.lazy(() => import('./DefaultFooter'));
 const DefaultHeader = React.lazy(() => import('./DefaultHeader'));
 
@@ -32,9 +30,16 @@ const VERSION_NAV_ITEM_ATTRS = {
     icon: 'cui-cog',
     url: 'https://rclone.org/changelog',
     variant: 'success'
-}
-class DefaultLayout extends Component {
+};
 
+// Компонент-обертка для использования хуков навигации
+const DefaultLayoutWithNavigation = (props) => {
+    const navigate = useNavigate();
+    const location = useLocation();
+    return <DefaultLayout {...props} navigate={navigate} location={location} />;
+};
+
+class DefaultLayout extends Component {
     loading = () => <div className="animated fadeIn pt-1 text-center">Loading...</div>;
 
     get navConfig() {
@@ -51,64 +56,74 @@ class DefaultLayout extends Component {
 
     componentDidMount() {
         if (!localStorage.getItem(AUTH_KEY) || window.location.href.indexOf(LOGIN_TOKEN) > 0) {
-            this.props.history.push('/login');
+            this.props.navigate('/login');
         } else {
             this.props.getVersion();
         }
     }
 
+    
+    signOut = (e) => {
+        e.preventDefault();
+        localStorage.removeItem(AUTH_KEY);
+        this.props.navigate('/login');
+    };
+
     render() {
-        // console.log("isConnected, default layout", this.props.isConnected);
         return (
-
-
             <div className="app" data-test="defaultLayout">
                 <ErrorBoundary>
-                    <AppHeader fixed>
+                    <CHeader fixed>
                         <Suspense fallback={this.loading()}>
-                            <DefaultHeader onLogout={e => this.signOut(e)}/>
+                            <DefaultHeader onLogout={this.signOut}/>
                         </Suspense>
-                    </AppHeader>
+                    </CHeader>
                     <div className="app-body">
-                        <AppSidebar fixed display="lg">
-                            <AppSidebarHeader/>
-                            <AppSidebarForm/>
+                        <CSidebar fixed visible={true}>
+                            <CSidebarHeader/>
                             <Suspense fallback={this.loading()}>
-                                <AppSidebarNav navConfig={this.navConfig} />
+                                <CSidebarNav items={this.navConfig.items} />
                             </Suspense>
-                            <AppSidebarFooter/>
-                            <AppSidebarMinimizer/>
-                        </AppSidebar>
+                            <CSidebarBrand/>
+                        </CSidebar>
                         <main className="main">
-                            <AppBreadcrumb appRoutes={routes}/>
+                            <CBreadcrumb className="d-md-down-none">
+                                {routes.map((route, idx) => route.name ? (
+                                    <CBreadcrumb.Item
+                                        key={idx}
+                                        active={this.props.location.pathname === route.path}
+                                        to={route.path}
+                                    >
+                                        {route.name}
+                                    </CBreadcrumb.Item>
+                                ) : null)}
+                            </CBreadcrumb>
                             <Container fluid>
                                 <Suspense fallback={this.loading()}>
-                                    <Switch>
-                                        {
-                                            routes.map((route, idx) => {
-                                                return route.component ? (
-                                                    <Route
-                                                        key={idx}
-                                                        path={route.path}
-                                                        exact={route.exact}
-                                                        name={route.name}
-                                                        render={props => (
-                                                            <route.component {...props} />
-                                                        )}/>
-                                                ) : (null);
-                                            })
-                                        }
-                                        <Redirect from="/" to="/dashboard"/>
-                                    </Switch>
+                                    <Routes>
+                                        {routes.map((route, idx) => {
+                                            return route.element ? (
+                                                <Route
+                                                    key={idx}
+                                                    path={route.path}
+                                                    element={<route.element />}
+                                                />
+                                            ) : null;
+                                        })}
+                                        <Route
+                                            path="/"
+                                            element={<Navigate to="/dashboard" replace />}
+                                        />
+                                    </Routes>
                                 </Suspense>
                             </Container>
                         </main>
                     </div>
-                    <AppFooter>
+                    <CFooter>
                         <Suspense fallback={this.loading()}>
                             <DefaultFooter/>
                         </Suspense>
-                    </AppFooter>
+                    </CFooter>
                 </ErrorBoundary>
             </div>
         );
@@ -120,4 +135,9 @@ const mapStateToProps = (state) => ({
     version: state.version,
 });
 
-export default connect(mapStateToProps, { getVersion })(DefaultLayout);
+const ConnectedDefaultLayout = connect(
+    mapStateToProps, 
+    { getVersion }
+)(DefaultLayout);
+
+export default DefaultLayoutWithNavigation;
