@@ -3,7 +3,7 @@ import {NavLink} from 'react-router-dom';
 import {Nav, NavItem} from 'reactstrap';
 import PropTypes from 'prop-types';
 
-import {AppNavbarBrand, AppSidebarToggler} from '@coreui/react';
+import {CSidebarBrand, CSidebarToggler} from '@coreui/react';
 import logo from '../../assets/img/brand/logo.png'
 import favicon from '../../assets/img/brand/favicon.png'
 import BackendStatusCard from "../../views/Base/BackendStatusCard/BackendStatusCard";
@@ -22,12 +22,12 @@ class DefaultHeader extends Component {
 
         return (
             <React.Fragment>
-                <AppSidebarToggler className="d-lg-none" display="md" mobile/>
-                <AppNavbarBrand
+                <CSidebarToggler className="d-lg-none" display="md" mobile/>
+                <CSidebarBrand
                     full={{src: logo, width: 89, height: 25, alt: 'Rclone Logo'}}
                     minimized={{src: favicon, width: 30, height: 30, alt: 'Rclone Logo'}}
                 />
-                <AppSidebarToggler className="d-md-down-none" display="lg"/>
+                <CSidebarToggler className="d-md-down-none" display="lg"/>
 
                 <Nav className="d-md-down-none" navbar>
                     <NavItem className="px-3">
