@@ -1,7 +1,7 @@
 import React from "react";
 import axiosInstance from "../../../utils/API/API";
 import {Alert, Col, Container, Row, Spinner, Table} from "reactstrap";
-import {DropTarget} from "react-dnd";
+import {useDrop} from 'react-dnd';
 import FileComponent from "./FileComponent";
 import {ItemTypes} from "./Constants";
 import {toast} from "react-toastify";
@@ -25,82 +25,49 @@ import ErrorBoundary from "../../../ErrorHandling/ErrorBoundary";
 import {createNewPublicLink, deleteFile, purgeDir} from "rclone-api";
 import {createSelector} from "reselect";
 import DropOverlay from "../../Base/DropOverlay/DropOverlay";
-/*
-* Start code for react DND
-* */
 
-const filesTarget = {
-    drop(props, monitor, component) {
-        if (monitor.didDrop()) return;
-        // console.log("drop", props, monitor, monitor.getItem(), component);
+// FilesViewWithDnD компонент для обработки drag and drop
+const FilesViewWithDnD = (props) => {
+    const [{ isOver, canDrop }, drop] = useDrop(() => ({
+        accept: ItemTypes.FILECOMPONENT,
+        drop: (item, monitor) => {
+            if (monitor.didDrop()) return;
 
-        let {Name, Path, IsDir, remoteName} = monitor.getItem();
+            const { Name, Path, IsDir, remoteName } = item;
+            const srcRemoteName = addColonAtLast(remoteName);
+            const srcRemotePath = Path;
+            const destRemoteName = addColonAtLast(props.currentPath.remoteName);
+            const destRemotePath = props.currentPath.remotePath;
 
-        let srcRemoteName = addColonAtLast(remoteName);
-        let srcRemotePath = Path;
-        let destRemoteName = addColonAtLast(props.currentPath.remoteName);
-        let destRemotePath = props.currentPath.remotePath;
+            return {
+                srcRemoteName,
+                srcRemotePath,
+                destRemoteName,
+                destRemotePath,
+                Name,
+                IsDir,
+                updateHandler: props.updateHandler
+            };
+        },
+        canDrop: (item) => {
+            const { remoteName, remotePath } = item;
+            const destRemoteName = props.currentPath.remoteName;
+            const destRemotePath = props.currentPath.remotePath;
+            if (destRemoteName === remoteName) {
+                return destRemotePath !== remotePath;
+            }
+            return true;
+        },
+        collect: (monitor) => ({
+            isOver: monitor.isOver(),
+            canDrop: monitor.canDrop(),
+        }),
+    }));
 
-        // console.log("drop:this", this);
-
-        return {
-            srcRemoteName,
-            srcRemotePath,
-            destRemoteName,
-            destRemotePath,
-            Name,
-            IsDir,
-            updateHandler: component.updateHandler
-        }
-
-    },
-    canDrop(props, monitor) {
-        const {remoteName, remotePath} = monitor.getItem();
-        const destRemoteName = props.currentPath.remoteName;
-        const destRemotePath = props.currentPath.remotePath;
-        if (destRemoteName === remoteName) {
-            return destRemotePath !== remotePath;
-        }
-        return true;
-    }
+    return <FilesView {...props} dropRef={drop} isOver={isOver} canDrop={canDrop} />;
 };
 
-function collect(connect, monitor) {
-    return {
-        connectDropTarget: connect.dropTarget(),
-        isOver: monitor.isOver(),
-        canDrop: monitor.canDrop()
-    }
-}
-
-
-/*
-* END code for react DND
-* */
-
-// Provides the up button view in the files view
-// function UpButtonComponent({upButtonHandle, gridMode}) {
-//     if (gridMode === "card") {
-//         return (
-//             <Col lg={12}>
-//                 <Button onClick={() => upButtonHandle()}>Go Up</Button>
-//             </Col>
-//         )
-//     } else {
-//         return (
-//             <tr onClick={() => upButtonHandle()} className={"pointer-cursor"}>
-//                 <td colSpan={1}/>
-//                 <td colSpan={4}><i className={"fa fa-file-o"}/> Go Up...</td>
-//             </tr>);
-//     }
-// }
-
-/**
- * FilesView component renders files in the file explorer.
- */
 class FilesView extends React.PureComponent {
-
-
     constructor(props) {
         super(props);
         this.state = {
@@ -110,28 +77,23 @@ class FilesView extends React.PureComponent {
             shouldUpdate: true,
             showLinkShareModal: false,
             generatedLink: "",
-
-
         };
         this.handleFileClick = this.handleFileClick.bind(this);
         this.downloadHandle = this.downloadHandle.bind(this);
         this.deleteHandle = this.deleteHandle.bind(this);
-        console.log("FilesView constructor", this.props);
     }
 
     closeLinkShareModal = () => {
         this.setState({
             showLinkShareModal: false
-        })
+        });
     };
 
     showLinkShareModal = () => {
         this.setState({
             showLinkShareModal: true
-
-        })
+        });
     };
-
 
     handleFileClick(e, item) {
         const {Path, IsDir, IsBucket} = item;
@@ -140,309 +102,253 @@ class FilesView extends React.PureComponent {
         } else {
             this.downloadHandle(item);
         }
-
     }
 
     updateRemotePath(newRemotePath, IsDir, IsBucket) {
         const {remoteName} = this.props.currentPath;
-
         let updateRemoteName = "";
         let updateRemotePath = "";
-
 
         if (IsBucket) {
             updateRemoteName = addColonAtLast(remoteName) + newRemotePath;
             updateRemotePath = "";
-            // backStack.push({remoteName: addColonAtLast(backStack.peek().remoteName) + remotePath, remotePath: ""});
-
         } else if (IsDir) {
             updateRemoteName = remoteName;
             updateRemotePath = newRemotePath;
-            // backStack.push({remoteName: backStack.peek().remoteName, remotePath: remotePath});
         }
         this.props.changePath(this.props.containerID, updateRemoteName, updateRemotePath);
     }
 
-
     getFilesList() {
         const {remoteName, remotePath} = this.props.currentPath;
-
         this.props.getFiles(remoteName, remotePath);
-
     }
 
     async downloadHandle(item) {
-        // let {remoteName, remotePath} = this.props;
-        let {remoteName, remotePath} = this.props.currentPath;
+        const {remoteName, remotePath} = this.props.currentPath;
         const {fsInfo} = this.props;
         let downloadUrl = "";
+        
         if (fsInfo.Features.BucketBased) {
             downloadUrl = `/[${remoteName}]/${remotePath}/${item.Name}`;
-
         } else {
             downloadUrl = `/[${remoteName}:${remotePath}]/${item.Name}`;
         }
 
-        this.setState((prevState) => {
-            return {
-                downloadingItems: prevState.downloadingItems + 1,
-                isDownloadProgress: true
-            };
-        });
+        this.setState((prevState) => ({
+            downloadingItems: prevState.downloadingItems + 1,
+            isDownloadProgress: true
+        }));
 
-        let response = await axiosInstance({
-            url: downloadUrl,
-            method: 'GET',
-            responseType: 'blob',
-        });
+        try {
+            const response = await axiosInstance({
+                url: downloadUrl,
+                method: 'GET',
+                responseType: 'blob',
+            });
 
-        const url = window.URL.createObjectURL(new Blob([response.data]));
-        const link = document.createElement('a');
-        link.href = url;
-        link.setAttribute('download', item.Name);
-        document.body.appendChild(link);
-        link.click();
+            const url = window.URL.createObjectURL(new Blob([response.data]));
+            const link = document.createElement('a');
+            link.href = url;
+            link.setAttribute('download', item.Name);
+            document.body.appendChild(link);
+            link.click();
+            link.remove();
+            window.URL.revokeObjectURL(url);
+        } catch (error) {
+            toast.error(`Error downloading file: ${error.message}`);
+        }
 
-        this.setState((prevState) => {
-            return {
-                downloadingItems: prevState.downloadingItems - 1,
-            };
-        }, () => {
-            if (this.state.downloadingItems === 0) {
-                this.setState({isDownloadProgress: false})
-            }
-        });
+        this.setState((prevState) => ({
+            downloadingItems: prevState.downloadingItems - 1,
+            isDownloadProgress: prevState.downloadingItems === 1 ? false : true
+        }));
     }
 
     async deleteHandle(item) {
-        let {remoteName} = this.props.currentPath;
+        const {remoteName} = this.props.currentPath;
 
         try {
             if (item.IsDir) {
-
                 await purgeDir(remoteName, item.Path);
-
-                this.updateHandler();
                 toast.info(`${item.Name} deleted.`);
-
             } else {
-
                 await deleteFile(remoteName, item.Path);
-                this.updateHandler();
-                toast.info(`${item.Name} deleted.`, {
-                    autoClose: true
-                });
+                toast.info(`${item.Name} deleted.`);
             }
-        } catch (e) {
-            // console.log(`Error in deleting file`);
-            toast.error(`Error deleting file. ${e}`, {
+            this.updateHandler();
+        } catch (error) {
+            toast.error(`Error deleting file: ${error.message}`, {
                 autoClose: false
             });
         }
-
     }
 
     updateHandler = () => {
-
-        // const {remoteName, remotePath} = this.props.currentPath;
         this.getFilesList();
     };
 
-    dismissAlert = (_) => {
+    dismissAlert = () => {
         this.setState({isDownloadProgress: false});
     };
 
-    linkShareHandle = (item) => {
+    linkShareHandle = async (item) => {
         const {fsInfo} = this.props;
-        if (fsInfo.Features.PublicLink) {
-            // console.log("Sharing link" + item.Name);
-            const {remoteName} = this.props.currentPath;
-            createNewPublicLink(remoteName, item.Path)
-                .then((res) => {
-                    // console.log("Public Link: " + res.data.url);
-
-                    this.setState({
-                        generatedLink: res.url,
-                        showLinkShareModal: true
-                    })
-                }, (error) => {
-                    toast.error("Error Generating link: " + error)
-                })
-        } else {
+        if (!fsInfo.Features.PublicLink) {
             toast.error("This remote does not support public link");
+            return;
         }
 
+        try {
+            const {remoteName} = this.props.currentPath;
+            const response = await createNewPublicLink(remoteName, item.Path);
+            this.setState({
+                generatedLink: response.url,
+                showLinkShareModal: true
+            });
+        } catch (error) {
+            toast.error(`Error generating link: ${error.message}`);
+        }
     };
 
     getFileComponents = (isDir) => {
         const {files, containerID, gridMode, fsInfo, loadImages} = this.props;
         const {remoteName, remotePath} = this.props.currentPath;
-        // console.log(fsInfo, files);
-        if (fsInfo && !isEmpty(fsInfo)) {
-            return files.reduce((result, item) => {
-                let {ID, Name} = item;
-                // Using fallback as fileName when the ID is not available (especially for local file system)
-                if (ID === undefined) {
-                    ID = Name;
-                }
-                if (item.IsDir === isDir) {
-                    result.push(
-                        <FileComponent key={ID} item={item} clickHandler={this.handleFileClick}
-                                       downloadHandle={this.downloadHandle} deleteHandle={this.deleteHandle}
-                                       remoteName={remoteName} remotePath={remotePath} gridMode={gridMode}
-                                       containerID={containerID}
-                                       linkShareHandle={this.linkShareHandle}
-                                       loadImages={loadImages}
-                                       isBucketBased={fsInfo.Features.BucketBased}
-                                       canCopy={fsInfo.Features.Copy} canMove={fsInfo.Features.Move} itemIdx={1}
-                                       >
 
-                        </FileComponent>
-                    );
-                }
-                return result;
-            }, []);
-        }
+        if (!fsInfo || isEmpty(fsInfo)) return [];
+
+        return files.reduce((result, item) => {
+            let {ID, Name} = item;
+            if (ID === undefined) {
+                ID = Name;
+            }
+            if (item.IsDir === isDir) {
+                result.push(
+                    <FileComponent 
+                        key={ID} 
+                        item={item} 
+                        clickHandler={this.handleFileClick}
+                        downloadHandle={this.downloadHandle} 
+                        deleteHandle={this.deleteHandle}
+                        remoteName={remoteName} 
+                        remotePath={remotePath} 
+                        gridMode={gridMode}
+                        containerID={containerID}
+                        linkShareHandle={this.linkShareHandle}
+                        loadImages={loadImages}
+                        isBucketBased={fsInfo.Features.BucketBased}
+                        canCopy={fsInfo.Features.Copy} 
+                        canMove={fsInfo.Features.Move} 
+                        itemIdx={1}
+                    />
+                );
+            }
+            return result;
+        }, []);
     };
 
     applySortFilter = (sortFilter) => {
-        const {changeSortFilter, containerID} = this.props;
-
-        if (this.props.sortFilter === sortFilter) {
-            return changeSortFilter(containerID, sortFilter, (this.props.sortFilterAscending !== true));
-        } else {
-            return changeSortFilter(containerID, sortFilter, true);
-        }
-
+        const {changeSortFilter, containerID, sortFilterAscending} = this.props;
+        return changeSortFilter(containerID, sortFilter, 
+            this.props.sortFilter === sortFilter ? !sortFilterAscending : true);
     };
-
 
     render() {
         const {isLoading, isDownloadProgress, downloadingItems, generatedLink, showLinkShareModal} = this.state;
-        const {connectDropTarget, isOver, files, gridMode, canDrop, sortFilter, sortFilterAscending} = this.props;
-        const {remoteName} = this.props.currentPath;
+        const {dropRef, isOver, files, gridMode, canDrop, sortFilter, sortFilterAscending, remoteName} = this.props;
 
         if (isLoading || !files) {
-            return (<div><Spinner color="primary"/> Loading</div>);
+            return <div><Spinner color="primary"/> Loading</div>;
+        }
+
+        if (remoteName === "") {
+            return <div>No remote is selected. Select a remote from above to show files.</div>;
+        }
+
+        const dirComponents = this.getFileComponents(true);
+        const fileComponents = this.getFileComponents(false);
+
+        let content;
+        if (gridMode === "card") {
+            content = (
+                <Container fluid={true}>
+                    <Row>
+                        <Col lg={3}>
+                            <h3>Directories</h3>
+                            <ScrollableDiv height={FILES_VIEW_HEIGHT}>
+                                {dirComponents}
+                            </ScrollableDiv>
+                        </Col>
+                        <Col lg={9}>
+                            <h3>Files</h3>
+                            <ScrollableDiv height={FILES_VIEW_HEIGHT}>
+                                <Row>
+                                    {fileComponents}
+                                </Row>
+                            </ScrollableDiv>
+                        </Col>
+                    </Row>
+                </Container>
+            );
         } else {
-
-
-            if (remoteName === "") {
-                return (<div>No remote is selected. Select a remote from above to show files.</div>);
-            }
-
-
-            let dirComponentMap = this.getFileComponents(true);
-
-            let fileComponentMap = this.getFileComponents(false);
-
-            let renderElement = "";
-
-            if (gridMode === "card") {
-
-                renderElement = (
-
-                    <Container fluid={true}>
-
-                        <Row>
-                            <Col lg={3}>
-                                <h3>Directories</h3>
-                                <ScrollableDiv height={FILES_VIEW_HEIGHT}>
-                                    {dirComponentMap}
-                                </ScrollableDiv>
-                            </Col>
-                            <Col lg={9}>
-                                <h3>Files</h3>
-                                <ScrollableDiv height={FILES_VIEW_HEIGHT}>
-                                    <Row>
-                                        {fileComponentMap}
-                                    </Row>
-                                </ScrollableDiv>
-                            </Col>
-
-                        </Row>
-
-
-                    </Container>
-
-                )
-            } else {
-                let filterIconClass = "fa fa-lg fa-arrow-down";
-                if(sortFilterAscending){
-                    filterIconClass = "fa fa-lg fa-arrow-up";
-                }
-                renderElement = (
-
-                    <Container fluid={true} className={"p-0"}>
-
-                        <ScrollableDiv height={FILES_VIEW_HEIGHT}>
-
-                            <Table className="table table-responsive-sm table-striped table-fix-head">
-                                <thead>
+            const filterIconClass = sortFilterAscending ? "fa fa-lg fa-arrow-up" : "fa fa-lg fa-arrow-down";
+            content = (
+                <Container fluid={true} className="p-0">
+                    <ScrollableDiv height={FILES_VIEW_HEIGHT}>
+                        <Table className="table table-responsive-sm table-striped table-fix-head">
+                            <thead>
                                 <tr>
-                                    <th className="pointer-cursor"
-                                        onClick={() => this.applySortFilter("name")}>Name {sortFilter === "name" &&
-                                    <i className={filterIconClass}/>}</th>
-                                    <th className="pointer-cursor"
-                                        onClick={() => this.applySortFilter("size")}>Size {sortFilter === "size" &&
-                                    <i className={filterIconClass}/>}</th>
-                                    <th className="d-none d-md-table-cell pointer-cursor"
-                                        onClick={() => this.applySortFilter("modified")}>Modified {sortFilter === "modified" &&
-                                    <i className={filterIconClass}/>}</th>
+                                    <th className="pointer-cursor" onClick={() => this.applySortFilter("name")}>
+                                        Name {sortFilter === "name" && <i className={filterIconClass}/>}
+                                    </th>
+                                    <th className="pointer-cursor" onClick={() => this.applySortFilter("size")}>
+                                        Size {sortFilter === "size" && <i className={filterIconClass}/>}
+                                    </th>
+                                    <th className="d-none d-md-table-cell pointer-cursor" onClick={() => this.applySortFilter("modified")}>
+                                        Modified {sortFilter === "modified" && <i className={filterIconClass}/>}
+                                    </th>
                                     <th>Actions</th>
                                     <th>Status</th>
                                 </tr>
-                                </thead>
-                                <tbody>
+                            </thead>
+                            <tbody>
                                 {files.length > 0 ? (
-                                        <React.Fragment>
-                                            <tr>
-                                                <th colSpan={5}>Directories</th>
-                                            </tr>
-                                            {dirComponentMap}
-                                            <tr>
-                                                <th colSpan={5}>Files</th>
-                                            </tr>
-                                            {fileComponentMap}
-                                        </React.Fragment>
-                                    ) :
-                                    <tr>
-                                        <th colSpan={5}>Files</th>
-                                    </tr>
-                                }
-                                </tbody>
-                            </Table>
-                        </ScrollableDiv>
-                    </Container>
-
-
-                );
-            }
-
-
-            return connectDropTarget(
-                <div className={"row"}>
-                    {isOver && canDrop && <DropOverlay/>}
-                    <ErrorBoundary>
-
-                        <Alert color="info" isOpen={isDownloadProgress} toggle={this.dismissAlert} sm={12}
-                               lg={12}>
-                            Downloading {downloadingItems} file(s). Please wait.
-                        </Alert>
-
-                        {renderElement}
-
-                        <LinkShareModal closeModal={this.closeLinkShareModal} isVisible={showLinkShareModal}
-                                        linkUrl={generatedLink}/>
-                    </ErrorBoundary>
-                </div>
+                                    <>
+                                        <tr><th colSpan={5}>Directories</th></tr>
+                                        {dirComponents}
+                                        <tr><th colSpan={5}>Files</th></tr>
+                                        {fileComponents}
+                                    </>
+                                ) : (
+                                    <tr><th colSpan={5}>Files</th></tr>
+                                )}
+                            </tbody>
+                        </Table>
+                    </ScrollableDiv>
+                </Container>
             );
         }
-    }
 
+        return (
+            <div className="row" ref={dropRef}>
+                {isOver && canDrop && <DropOverlay/>}
+                <ErrorBoundary>
+                    <Alert color="info" isOpen={isDownloadProgress} toggle={this.dismissAlert} sm={12} lg={12}>
+                        Downloading {downloadingItems} file(s). Please wait.
+                    </Alert>
+                    {content}
+                    <LinkShareModal 
+                        closeModal={this.closeLinkShareModal} 
+                        isVisible={showLinkShareModal}
+                        linkUrl={generatedLink}
+                    />
+                </ErrorBoundary>
+            </div>
+        );
+    }
 }
 
-const propTypes = {
+FilesView.propTypes = {
     containerID: PropTypes.string.isRequired,
     currentPath: PROP_CURRENT_PATH.isRequired,
     fsInfo: PROP_FS_INFO,
@@ -451,38 +357,33 @@ const propTypes = {
     loadImages: PropTypes.bool.isRequired
 };
 
-const defaultProps = {};
-
-
-FilesView.propTypes = propTypes;
-FilesView.defaultProps = defaultProps;
-
+FilesView.defaultProps = {};
 
 const getVisibleFiles = createSelector(
-    (state, props) => props.containerID,
-    (state, props) => state.explorer.currentPaths[props.containerID],
-    (state, props) => state.explorer.visibilityFilters[props.containerID],
-    (state, props) => state.explorer.sortFilters[props.containerID],
-    (state, props) => state.explorer.searchQueries[props.containerID],
-    (state, props) => state.explorer.sortFiltersAscending[props.containerID],
-    (state, props) => state.remote.files[`${state.explorer.currentPaths[props.containerID].remoteName}-${state.explorer.currentPaths[props.containerID].remotePath}`],
+    [
+        (state, props) => props.containerID,
+        (state, props) => state.explorer.currentPaths[props.containerID],
+        (state, props) => state.explorer.visibilityFilters[props.containerID],
+        (state, props) => state.explorer.sortFilters[props.containerID],
+        (state, props) => state.explorer.searchQueries[props.containerID],
+        (state, props) => state.explorer.sortFiltersAscending[props.containerID],
+        (state, props) => state.remote.files[`${state.explorer.currentPaths[props.containerID].remoteName}-${state.explorer.currentPaths[props.containerID].remotePath}`],
+    ],
     (containerID, currentPath, visibilityFilter, sortFilter, searchQuery, sortFilterAscending, files) => {
         files = files.files;
-        // Filter according to visibility filters
+        
         if (visibilityFilter && visibilityFilter !== "") {
             files = changeListVisibility(files, visibilityFilter);
         }
 
-        //Filter according to search query, if any
         if (searchQuery) {
             files = changeSearchFilter(files, searchQuery);
         }
-        files.sort(getSortCompareFunction(sortFilter, sortFilterAscending));
 
+        files.sort(getSortCompareFunction(sortFilter, sortFilterAscending));
         return files;
     }
-)
-
+);
 const mapStateToProps = (state, ownProps) => {
     const {currentPaths, gridMode, searchQueries, loadImages, sortFilters, sortFiltersAscending} = state.explorer;
     const {containerID} = ownProps;
@@ -492,25 +393,23 @@ const mapStateToProps = (state, ownProps) => {
     const mloadImages = loadImages[containerID];
     const sortFilter = sortFilters[containerID];
     const sortFilterAscending = sortFiltersAscending[containerID];
-
+ 
     let fsInfo = {};
     const {remoteName, remotePath} = currentPath;
-
+ 
     if (currentPath && state.remote.configs) {
         const tempRemoteName = remoteName.split(':')[0];
         if (state.remote.configs[tempRemoteName])
             fsInfo = state.remote.configs[tempRemoteName];
     }
-
+ 
     const pathKey = `${remoteName}-${remotePath}`;
-
     let files = state.remote.files[pathKey];
-
+ 
     if (files) {
         files = getVisibleFiles(state, ownProps);
     }
-
-    // Sort the files
+ 
     return {
         files,
         currentPath,
@@ -521,12 +420,9 @@ const mapStateToProps = (state, ownProps) => {
         sortFilter,
         sortFilterAscending
     }
-};
-
-export default compose(
-    connect(
-        mapStateToProps, {getFiles, navigateUp, changePath, changeSortFilter}
-    ),
-    DropTarget(ItemTypes.FILECOMPONENT, filesTarget, collect)
-)(FilesView)
-
+ };
+ 
+ export default compose(
+    connect(mapStateToProps, {getFiles, navigateUp, changePath, changeSortFilter}),
+    // Оборачиваем компонент в DnD
+ )(FilesViewWithDnD);
